Customize the Form Assembly e-Signature message about verifying email. (#647)
* Customize the Form Assembly e-Signature message about verifying email.

The built-in behavior for Form Assembly e-Signature forms is that on the
final page where you sign, it shows a warning message saying:

Incomplete Response
Please click the link in the verification email to complete your signature.

However, you don't get the email until after you click the button to submit.
This is confusing and we have no way to control this on the Form Assembly
side of things. So this commit uses JS to change the language to be more clear.

* This happens on page load. No need to explicitly send a beacon.

* Try and stick with our Honeycomb field naming conventions.

Note: I'm going to followup with a separate PR to make the "controller"
and "action" (aka the filename and function) that we construct the
HoneycombSpan with more normalized when things make it to Honeycomb.

E.g. things like the following

name = js.page.load
js.app.controller = form_assembly
js.app.action = changeESignatureMessage
js.app.name = js.form_assembly.changeESignatureMessage
js.app.form_assembly.esignature_message_found = true

* Add basic Jest test. Fixed bug it found.

* Add test for running on page with no e-Sig warning message

* Merge with: https://github.com/bebraven/platform/pull/649

* snake_case to CamelCase in JS