Normalize the Honeycomb span names to be snake_case. (#660)
The guidance on field naming is all snake_case. Let's be consistent
with the names of the spans too so that when querying in Honeycomb
you can always start typing stuff in snake_case without thinking if you're
looking for something.

Note: the ONE exception is the functionName of the Javascript spans
that stick with CamelCase. Those functions do not map to the filename,
which the Ruby spans will since we have generally one class per file,
so you want something greppable when looking at a span and trying to
figure out where in the code it is.