Add Project autosave indicators (#675)
* Add Project autosave indicators

* Clarify padding with comment

* First pass at CSS styling.

Puts the status bar inside the container so that full width relative
to the parent is inline with the content, uses Braven brand colors
for the background and text, reduce padding since this is supposed to
be a small indication at the bottom in the normal case (will add more
padding most likely for the failure cases), make the font slightly
small than the body text for the same reason - this should be out
of the way as much as possible. It's really when things fail that
we want it more prominent.

* Add classes to be able to style success and failure different.

* Style error status bar differently than success

* Have the Retry links on failed answers retry them all.

Most likely if you are working on an answer and it fails to save,
you'll stop and try to figure out why before continuing. But some
folks just keep working (imagine you're rushing to get the project
done before the deadline coming up in 30 min). It's annoying to go
through one-by-one and click Retry once you have internet connection
again, so just have the retry on any answer do it for all unsaved
answers.

* Make the the 'Saving...' message visibly flash.

It happens so fast normally that you can't actually see the words, so
give it a split second before setting success messages after save completes

* Only send answers when they change.

With the new status indicators, especially in the case of failure,
just clicking around into fields even if you don't actually type
anything or make any changes will cause it to show up as a
"failed to save" answer in red.

This change just stores the value of the input field on a new variable
attached to the element and then only sends the answer on blur if it changes.

* whitespace

* Add a third styling state for 'Saving ...' so it actually draws attention

When I was doing this on a small screen the "All answers saved" status
bar was within my scope of vision. On a bigger screen it's not when you're
working on the top stuff. This change flashes the "Saving..." status a
bit more prominently so it draws your eye's attention.

* Force send answers on retry.

The new "only send changed" answer logic needs to overridden when
retrying failures

* Use "bold" instead of 700 for font weight

* Remove extra code

* Switch logic to only send changes to use JS event system.

* Switch behavior to a fade out of the "All progress saved" bar.

This makes it so that we don't have to purposefully make the
'Saving...' status more prominent and last longer. Instead, just
let that happen super fast (and screenreaders won't usually even
have to read it) and fade out the "All progress saved" status bar
which provides enough visual cues to catch your eye.

Note on code comment about the long fade out in the code to make
it more likely that a screenreader will get to it.

* Disable Submit button when there are failures to save answers.

Pretty hacky and inspiration taken from here:
https://stackoverflow.com/questions/31856712/update-component-state-from-outside-react-on-server-response

I don't know React (yet) and we have plans to cutover the whole Project stuff
to React at some point, so this is what I came up with...?

* Remove unused code.

* Incorporate Product team style change requests.

The "Saving..." status bar was still a little too light and Product
didn't think folks would notice it. Since it fades out, they didn't
think having it a bit more prominent was intrusive / distracting.
The feeling is that after the first few times it happens, you will
just forget about it (like living next to an elevated train).

Also, when there are errors add more language about the fact that they
should stop working and make sure they can get their answers to save.

* More actionable message for "some failed" message too.

Co-authored-by: Brian Sadler <sadler.brian@gmail.com>