Automatically change login email in Platform and Portal when changed in Salesforce (#677)
* Add `change_email` API endpoint that will be called from Salesforce.

* Turn of CSRF.

The previous iteration did this, but still generated logs about not
being able to verify it and was a little less obvious from the code.

* Add policy

* Change the route to be more generic "update_contact"

Right now, this is only run using Process Builder in Salesforce
that is triggered by changes to the email field. But, it sends
across the whole Contact from SF, so by making this more generic
we'll be able to trigger on other fields, like First Name or Last Name
and not have to go change the Salesforce code or refactor this endpoint
(maybe just add another field or two that we parse out)

* Implement the basic functionality. This works end to end now.

Error handling and specs still needed.

* Clarify comments

* Add Honeycomb instrumentation

* Add TODO

* Change the Notification emails too in Canvas

* Rollback local DB changes if Canvas changes fail

* Cleanup Honeycomb span. Don't need span at outermost level.

Just add the fields. Each action already has a span.

* Email sync failures to the Saleforce user that changed the email

* Better instrumentation

* Some cleanup

* Implement reconfirmable so they get a notification of the login change.

This ended up being a bit more involved than expected since when you
turn this setting on, the "email" column doesn't actually get updated.
Instead, the new email is put in the "unconfirmed_email" column until
they actually confirm. The flow shares the same confirmation email
as the original sign_up / registration, so the language has to be
vague enough for both scenarios.

* Move policy for existing Sync To LMS stuff to the new policy class

* Add Asana task to TODO since we'll do in a future PR

* Remove TODO and move to Asana task.

I think it's better to rename the Program Canvas Course related
fields to Canvas Cloud now vs naming this one with Highlander in it.
https://app.asana.com/0/1174274412967132/1200197329627699

* Don't send confirmation instructions twice.

The code was written to skip the confirmation instructions when
we update the user model to be registered IFF we're creating a
new record. This caused two emails to go out for existing users.
Wait until we successfully finish all the registration steps
(like creating their Canvas account) until we force the confirmation
email to go out (once).

* Change language from "click this link" to "click this button"

It's a button.