Allow Sync From Salesforce (aka Sync To LMS) to send account creation sign-up emails (#686)
* Consolidate sync_from_salesforce_program_form.html.erb with partial

The partial isn't used anywhere else. No need for it yet.

* Add checkbox to form allowing them to send sign-up emails.

Pass this down to the service.

* Placeholder sign-up email (so I can tinker and see diffs)

* Email preheader

* Left align instead of center (less spammy / marketing looking).

Also see followup to fix this in other emails:
https://app.asana.com/0/1174274412967132/1200239858551396

* Finish drafting the email

* Add method to send email on User model

* Actually send the sign_up email to newly created User records.

IFF the staff member indicated they should be sent.

* Prepopulate program_id and notification email if passed as params

* Show email preview and note about being able to confirm list first

* Tweak the HTML a bit.

* Add basic flow to allow confirming the list of users getting sign-up emails

* Refactor sample email into partial and show on confirmation page too

* Strip whitespace from user input

* Actually fetch the list of users that would get emails for display

* Delete '

Somehow got file named `'` in my root that was added...

* Comments

* Filter list to only Enrolled Participants.

This could happen if someone is marked Enrolled, is never synced,
then is changed to Dropped. The Participant would come through in the
list to sync but not have an account. We wouldn't create a user (I hope).

* Add TODO for small bug.

* Typo