Reconcile mismatched emails across Salesforce, Platform, and Canvas. (#695)
* Reconcile mismatched emails across Salesforce, Platform, and Canvas.

Salesforce is the source of truth for what the login email should be.
If it doesn't match what is in Platform and Canvas, we update it.
In order for the updated email to be active and able to login,
a reconfirmation email is sent out. This also serves to notify the
user since staff could change their login email without their knowledge.

A big change here is that instead of updating the Canvas login email
at the time of sync'ing, we wait until they actually reconfirm it
by clicking the confirmation link in the email they get. Their old
account will continue to work until they activate the new one. This
was required so that we didn't continue to spam them with reconfirmation
emails if they don't confirm the change. We just do it once and if
they don't activate that email we skip trying to sync it.

* Fix to send failure email even if platform User isn't found

* Rollback confirmation failures due to Canvas API failures.

We run the logic to get Canvas setup properly after Devise considers
the confirmation complete. We want to let Devise do it's thing first
in case there is a fundamental issue with the confirmation, like a bad
or missing token. But if we fail to make the Canvas API calls, we still
want the confirmation link to work b/c presumably Canvas was down or
we deployed a bug that we should revert and once that situation is fixed
we want to be able to ask users to just try again.

* Add some notes about what specs to write for confirmation logic

* Fix specs and Admin User creation

* Better behavior for invalid confirmation_tokens

* Missing downcase when comparing emails. Also move Honeycomb fields

* Code review changes plus bug fix (unconfirmed_email can be nil)

* Fix specs