Don't overwrite manually entered grades (#697)
* Basic working implementation to skip auto-grading if the grade was manually set

* Add CanvasAPI specs

* Fix grade_module_for_user_job_spec

* Fix grade_modules_spec

* Manual grading specs for grade_module_for_user_job

* Switch "scenario" for "it" in spec

Copied scenario from another file when trying to get a shared_example to
work but I just read up that it's for feature specs and is the same as "it".

* Add grade_modules_specs for manually overridding grade

* Small tweak for readability

* Clean up comments and backlog TODO

* Address code review feedback

* Show warning message on Module submission if manually graded

Showing this to TAs, staff, as well as Fellows. We can revisit
in the future if Fellows shouldn't see it, but let's not add complexity now.