Consolidate CANVAS_URL and CANVAS_CLOUD_URL. (#715)
The old Portal code paths are no longer supported. Canvas is always
the cloud instance of Canvas. Cleanup that stuff a bit.

Also make them available as a global constant for things that aren't
trying to use the API and just need the URL (e.g. redirecting to the
Canvas login path).