Skip Participant.Status = nil in the nightly sync. (#717)
For InternPro, we're not giving LC's access to Canvas. We still want
to have them as Participants in Salesforce so that we can do Campaign Monitor
driving communication. Only a Status of 'Enrolled', 'Completed', or 'Dropped'
has sync logic associated with it anyway, so just skip Participants with no status.
This will prevent them from getting User records in Platform and accidentally
getting emails or somehow ending up in Canvas.

https://app.asana.com/0/1174274412967132/1200311875342975