Specs and bug fix for user needing reconfirmation for changed email. (#727)
* Specs and bug fix for user needing reconfirmation for changed email.

This started off as writing specs for a user whose email changes
and they need to confirm their new email to be able to log in with it:
https://app.asana.com/0/1174274412967132/1200325250500042

I realized that the behavior we had talked about here:
https://github.com/bebraven/platform/pull/685#discussion_r625961906

was never implemented. So in addition to the specs, it implements
that behavior so that users trying to log in with their new email
but who haven't confirmed can now generate a new confirmation email
if they lost it or the token expired.

* DANGER comments

* Change login instructions to clarify what happens if they don't press the button

If staff chagnes it to an incorrect email or in the future folks can change it themselves (and try to get into someone elses account that way) we want to make it clear that nothing will happen until they click the button and that if the email wasn't intended for them to just ignore it.

The language needs to make sense in the context of a registration confirmation vs an email change confirmation.

* Code review feedback.

I did a little more than what was talked about in the code review by
actually rewriting the BravenCAS:CustomAuthenticator to make its usage
harder to call incorrectly. All three methods now just take `credentials`
and I added specs to make sure I did it right.