Don’t stay on the login form if you actually login (#733)
* If no service parameter is passed to the login, use a default to redirect.

Prior to this, in certain scenarios (e.g. if you use a bad confirmation link
but then click "try logging in instead"), it would log you in and say
"You have successfully logged in" but stay on teh login page. Now, if you're
and admin it goes into platform and if not it redirects to Canvas.

* Don't auto-login users on password reset.

We had changed this behavior in this PR:
https://github.com/bebraven/platform/pull/220

But subsequently broke it when we subclassed the passwords controller.
This brings back the old behavior. The reason is b/c we want them to
immediately be forced to use their password which will hopefully cause
them to either update their password manager or write it down however
they manage passwords. Basically, don't incentivize using password
reset every time to log in.

* Make confirmations auto-login share code with new "default service"

TODO: haven't tested this yet.

* Make "resend instructions" on the "PLEASE CONFIRM" page indicate that the button was pressed.

We don't want to expose whether the UUID used to get to:
  https://platformweb/users/registration?login_attempt=true&uuid=something
is valid, but we do need to give an indication that the button has been pressed.
This is just the https://platformweb/users/confirmation/new endpoint which
gives the same message but without the button to resend. This used to work
and was introduced here: https://github.com/bebraven/platform/pull/193
but subsequently broken here: https://github.com/bebraven/platform/pull/685

Note that you have to get the password to correct to get to the page where
you can resend confirmation instructions. This change isn't exposing anything
about the UUID b/c it always behaves the same.

* Specs for reset password not signing you in

* Specs for resending confirmation email