Add support for generating Zoom links (#743)
* Straight copy of all Zoom related files from staging.

https://github.com/bebraven/platform/pull/430

Need to fix this up and pull code out of rowan_bot gem
into this repo as well.

* Mostly implement "Generate Webinar From CSV" feature.

I'm going to commit the ZoomAPI changes to make it work separately.

Test Plan:
- Try with a .csv generated using the template
- Try with a .csv saved as "CSV UTF-8 (Comma delimited)" on a Mac
- Try with a .csv saved as "CSV (Comma delimited" on a Mac

* Refactor ZoomAPI to be more like CanvasAPI and work in Rails.

* Cleanup: rename Webinar to Zoom. Clarify language/instructions.

Also, rip out more of the Sync To Webinar stuff which I'll be cutting
over to be a part of the normal Salesforce sync

* Fix bug where we created Users that weren't Enrolled.

I'm starting to add the Zoom generation and sync logic to the
normal Program sync. Fixing the following bug:
https://app.asana.com/0/1174274412967132/1200239858551410

* Add Zoom fields, Id, and VolunteerRole to SFParticipant struct

This were also added in BZ_ProgramParticipantInfoService.apxc
in the Salesforce APEX repo and deployed to Salesforce Prod.

* First rough pass of syncing Zoom links in the SF sync.

Only implemented assigning links and updating salesforce. In addition
to implement unassigning when they drop, there are a lot of TODOs

* Refactor and add Honeycomb fields

* Implement unassigning Zoom links when they drop and fix bug

- When Dropped, their links should always be cleared out.
- When Enrolled, if the meeting is no longer configured on the Cohort Schedule
  we don't clear our their link like before this change. It had a bug where
  if there was no change it will accidentally clear out the link and the code
  gets really messy if we try and clear it out just b/c it's no longer on the
  Cohort Schedule.

* Implement prefixes

* Remove obsolete file.

* Add "Force Zoom Update" option.

* Fix comment

* Fix comment

* Add link to task for future TODO

* Better instructions for generating Zoom links using a .csv

* Add policies

* Generate JWTs to authenticate instead of using test one with expiration.

Note: this changes the ZoomAPI.client from a singleton to a factory
method that produces a client object with valid auth for 30 seconds.
After that time, you'll need to get a new one using the method.

* salesforce_controller_spec.rb

* sync_from_salesforce_program_job_spec.rb

* zoom_policy_spec.rb

* salesforce_controller_spec.rb - test that the force_zoom_update param gets passed through

* generate_zoom_links_spec.rb

* sync_portal_enrollments_for_program_spec.rb

* First pass of sync_zoom_links_for_participant_spec.rb

A lot more to do, but here is a skeleton.

* Test when links are already synced

* Second pass at sync_zoom_links_for_participant_spec.rb

Refactors in shared_examples for Enrolled folks and runs the tests
for Fellows, LCs, and CPs.

TODO: Dropped enrollment tests

* Implement Dropped tests for sync_zoom_links_for_participant_spec.rb

* Tests for forcing the links to be updated

* Add specs for parsing the .csv used to generate Zoom links

* salesforce_api_spec.rb - #update_zoom_links

* Honeycomb fields

* Fix typo in div. Missing ending `>` symbol

* Remove duplicate method. Bad merge.

* Add test for Google Sheets .csv