UI tweaks (#753)
* Different copy for 1-1's vs attendance when Fellow views the assignment

Also, add generic language about checking your syllabus for how it
impacts your grade. For some programs, attendance is not graded
(university requirement) while for others it is.

* Add padding to all pages in the lti_canvas layout.

The mobile app launches everything in a new window. We need all
lti launches to have a bit of padding.

Note that most of the pages are shown in a window of their own now and
need the padding even on desktop.

Pages that I double checked aren't impacted b/c we want them iframed with
no padding:
- rate this module
- linkedIn authorization

Everything else with this layout benefits from this.

* Change some attendance to be consistent.

This PR: https://github.com/bebraven/platform/pull/725 was intended
to left align the pages that are shown inside of Canvas to match the
"Launch External Tool" prompt, but somehow that's not taking effect.
Generally, we're not very consistent with the styling across all our pages,
and we should do a bigger project to fix that up, but for now at least
make these two pages consistent with the change from that PR so that
when we do fix it all up, these will turn up in a grep.

* Make the "Impact Survey Submitted" page consistent with others.

The Waivers and Pre/Post Accelerator survey use this layout / styling.
Do the same for Impact Surveys.

* Rename Capstone Evaluations assignment.

Note that Capstone evaluations are submitted by both Fellows, as a
review of their peers, and by LCs as a review of Fellows in their Cohort.