More honeycomb stuff (#764)
* Add 'app.user.email' field to Honeycomb for CAS login stuff.

Note: this obviously doesn't end up with the user email in the
query results for the GET of the login page, but after then enter it
all other CAS related actions will now show which user they were for.
Could be helpful for troubleshooting login issues.

* Prefix user context with the calling class name when instrumenting multiple users.

It's going to be easy to mess up the span vs trace stuff in the future and
accidentally overwrite all the spans with some single user's context in
the entire trace. Prefix the user fields with the calling class's name
to avoid that.