Javascript package upgrades (aka npm/yarn)  (#798)
* Upgrade rails/webpacker 5.2.1 -> 5.4.0

Aiming to get rid of dependency on old `meow` package that depends
on an old `trim-newlines` package with a security issue.

* Swap deprecated node-sass for new dart-sass (just called sass).

TODO: make sure nothing will break in our repo.

One article to start with: https://panjiachen.github.io/vue-element-admin-site/guide/advanced/sass.html#upgrade-plan

* Get rid of some warnings after the last upgrades: @babel/plugin-proposal-private-methods

* Downgrade dart-sass (aka sass) to v1.32.12

Tons of warnings about deprecated division operator.

See:
- https://github.com/silinternational/ui-components/pull/47
- https://github.com/twbs/bootstrap/issues/34051

* Run `yarn upgrade`

This upgrades all packages based on their version range. There was
only one left. Next up is to upgrade what we can to new major versions.

* Downgrade libhoney to get rid of Libhoney::TransmissionClient errors

I thought they were gone (the original one here is: https://app.asana.com/0/1174274412967132/1200055379602510/f)
but there are still tons randomly all the time.

* Upgrade rails/webpacker minor version.

* Upgrade to latest minor version of webpack.

* Upgrade sass-loader and style-loader to latest version that work with webpack@4

* Upgrade webpack-merge to latest version

* Upgrade css-loader to latest version compatible with webpack v4

* Change babel versions in package.json to latest upgraded versions used.

Makes it easier to see at a glance the last time we "installed" these
and checked all the functionality.

* Upgrade jest.

* Upgrade webpack-dev-server

* Change dev note about WEB_CONCURRENCY setting

* Fix jest specs

* Downgrade webpack-dev-server to v3x.

I upgraded it to get rid of a dependency on a package with a vulnerability,
but it seems to depend on a version of webpack-cli that is only officially
compatible with webpack v5, which we can't move to until we can upgrade React stuff.

* Downgrade style-loader to fix ckeditor

Co-authored-by: rshipp <github@rshipp.com>