Add more instructions to Waivers for finding e-sig email. (#801)
* Add more instructions to Waivers for finding e-sig email.

We have to have them find the email they get and click the link
to verify their signature. For folks that don't do it, we have to
follow-up and get them to complete that step or else we need to remove
their access from the course after a couple weeks. Anything we can
do to make it more likely they complete this step on their own reduces
work and noise.

* Fix specs