Make our links more accessible. (#803)
* Make our links more accessible.

Task: https://app.asana.com/0/1174274412967132/1200411617157219

The difference is barely noticeable b/c we're just going from a 2.85:1
ratio to a 3:06:1 ratio, but better to actually pass than almost pass.

Also, this came up as part of not noticing it on the page that asks
you to try logging in instead when you have a bad confirmation link.
Remove the bold that I randomly put in that sentence about maybe having
already confirmed your account. I think that solves the issue with that
page and not noticing the log-in link.

* Use underlines instead of color for accessibility.

The Rise360 content already uses underlines to make links accessible.
Let's be consistent.