Handle Zoom errors more gracefully. (#808)
* Handle Zoom errors more gracefully.

Two scenarios where we got 400 BadRequest ZoomAPI responses that
we now translate into a nicer error message:
- For meetings that have ended
- For meethings where we're trying to register the host

Note: for the situation where we're trying to register the host, don't
report that failure in the Sync Email. Instead, just return a "link"
with the message about why we can't create a link and report that the
Sync was successful. It's a valid situation to make the staff member
who is a host of the Zoom meeting also a Teaching Assistant Participant
in the Program.

* Specs

* Handle bad emails during Generate Zoom Links process.

This error handling applies to the Sync process as well, but we don't
really expect it there.

Note: needed to add teh 'Accept: json/application' header to the Zoom API
b/c for some reason, that particular error was returned as XML without it.