Set all Module submissions to be on Jan 1, 1970 (#840)
* Set all Module submissions to be on Jan 1, 1970

Task: https://app.asana.com/0/1174274412967132/1200954533308126

The Canvas UI will show "Late" in red if the submitted_at time is after
the due date. We create a single placeholder submission the first time
a Fellow opens the Module and use that to post auto-grades too. The
"Submitted At" time has no meaning in terms of their grade since it's
the timestamp of when they opened it. If they open it for the first time
after the due date, it shows "Late", otherwise is always shows on-time.
This can be confusing. With this change, "Late" will never be shown
and the non-sensical date of Jan 1, 1970 should make support easier
if Fellows needs help understanding their grade and why they didn't
receive on-time credit. It'll be easy to say "ignore that, it doesn't
mean anything" - the on-time credit is given when you reach 100% completion
before the due date, which is shown on the grades page.

Note: the proper fix for this would be to send a new submission to Canvas
everytime they do more work. You can't update just the submited_at time.
That would require a major refactor and may put unecessary load on the server
in terms of API calls and database queries in addition to handling all the
tricky logic to be able to send the proper grade everytime (for example,
we need the submission to determine if the grade was manually overridde
which we'd have to make an API call to get and we'd still have to handle
0'ing out grades when the due date passes if they've never opened it as
well as handling extensions as well as clicking around the Module which can
send new progression statements to lrs_xapi_mock). This solution seemed to
be the right balance of acheiving the intended behavior in a simple way.

* Fix specs and switch to subsecond precision for timestamps.