BUG: Floating point math issue with grade computation (#854)
* Fix floating point math bugs

Move computation of the grade inside ComputedGradeBreakdown so
that we make sure it matches the displayed component breakdown.
To do this, switch from sending a percent to Canvas and instead
send the raw score that we compute. Compute that raw score by
adding up the score for each component that we show. This way,
the display of the grade breakdown matches the total score it
sums up to which matches what we store in Canvas, therefore avoiding
re-computing b/c of rounding errors.

https://app.asana.com/0/1174274412967132/1201022023697611

* Switch hardcoded full credit to use the constant.

* Switch to use Time objects instead of Strings for completed_at

This is really just for the specs. We already used the created_at timestamp
on the Rise360ModuleInteraction which is an ActiveRecord::TimeWithZone object.
This makes the specs match that behavior. I tested this manually as well
by completing a module late and seeing that the displayed message about
"You completed this module late, at BLAH" was the correct local time