Ability to send confirmation email from Admin User dashboard (#861)
* Add new button to resend the confirmation email.

Note: this also refactors the :CanSendNewSignupEmail to a more
generic :CanSendAccountCreationEmails role. Need to run a migration
when deploying to prod.

* Fix specs

* Remove Register and Confirm buttons from User Admin dash.

Let's stick to the normal process as closely as possible and just
have support tools to resend the emails, but not buttons to force it
to happen. That requires too much underlying information about the
motivation and context for signup and confirmation and registered as
concepts.

* Write "Send Confirmation Email" specs

* Migrate CanSendNewSignUpEmail Role to CanSendAccountCreationEmails

We don't need to control the ability to send these emails individually.
Made the name more generic and this migrates the existing role to the
new name.

* Confirmation emails should only be sent to registered users.

Registration is the first step (aka creating their account with a
password). Don't let staff send confirmation emails to accounts
that haven't already registered.

* Add Asana task to TODO re: regenerating sign-up tokens.