Hotfix: Remove massive whitespace below Rate This Module (#867)
* Make clicking "Submit" on Rate This Module not feel like you're done

- The big red "Submit" button made it feel like you just submitted the
  entire Module, especially b/c ...
- There is a MASSIVE amount of whitespace below the iframe.

I'm not sure what changed to cause the whitespace b/c we tested this stuff.
I tested with iframes created in Rise360 that didn't have a height set explicitly.
We've been publishing them with a height of 750px though which I assume is what
was tested in the dry run. Either way, Rise360 adds an inline
style="padding-bottom: BIG NUM;" when a height is set and the value we set
didn't take precedence over the inline style (which I never saw b/c I didn't
set an explicit height when authoring). Just add !important.

* Revert button color / text changes.

With the whitespace gone, this *should* be the end of the Module
and we want to encourage them to submit the feedback. Let's see
if the issue with folks not scrolling is fixed without changing
the styling to make it NOT feel like you're submitting the Module.