BUG: Rate This Module CSS wasn't being applied if it was navigated to (vs initially loaded) (#869)
* Makes the Rate This Module CSS apply when navigating there.

Currently, the code only works when you load the Module and the bookmark
brings you back to the section (aka lesson) that has the iframe in it.
If you navigate there from previous sections, the JS doesn't run again
and the iframe is AJAX loaded.

This has a bug in it where if you go back to the Module Overview
after already having entered the module, then we lose the observer
responsible for adding the CSS. Punting on that for now b/c the plan is
to DRASTICALLY simplify this by just setting up the observer on the top
`<div id="app">` and looking for the iframe for every mutation. We
need to test on a crappy device to make sure there isn't performance concerns
though.

* Simplify RateThisModule CSS addition by observing whole DOM.

The prior commit tried to be smart and only observe the nodes we need
to but the code was WAY over complicated. Let's just observe the whole
DOM. I tested this on a phone and didn't see any performance issues just
clicking around. I think the simplicity is worth not worry about performance
unless it becomes a problem.

* Add the iframe_found field back.

This may be useful to setup an alert on if these stop coming through.

* Switch Honeycomb field from string to boolean