Mark a Module as 100% complete once Rate This Module iframe is loaded (#874)
* Mark a Module as 100% complete once Rate This Module iframe is loaded

This is the last thing in a Module and even though there is very little
space below it that the user has to scroll to in order for the Rise360
package to send a 100% progress xAPI statement just to be safe we're
going to save the progress as 100% manually so that their guarenteed to
get full engagement credit if they hit the big red Submit button on
Rate This Module and then close it.

* Refactor to have all progress interactions use the helper method.

It will make maintaince easier in case we ever change what we store
or actually implement the DB constraint. This also makes it so that
generally only a single progress value is recorded per user/assignment
which will reduce the size of the table and allow us to simplify the
grading code once we introduce the DB constraint. We can't simplify now
b/c there is still a race condition and dupes could be created, though
extremely unlikely.

* Specs

* Use implicit return. More canonical?

* Fix logic for determining the on-time grade.

Before this, if a progress came in after the completed_at
interaction and before the due date, it would use the lower
progress and say it wasn't completed on time. We didn't think
lower progress could come in when initially implementing, but it can.

* Specs to detect on-time bug

* Clarify comment.