SalesforceAPI: make `find_participant()` require a `program_id` (#885)
* Make find_participant(contact_id, program_id) require params.

Before this change, find_participant() allowed nil for the program_id
and would just arbitrarily return the first Participant of any Status
for any Program. The Contact and Program combo should uniquely map to
a single LC, Fellow, or TA and we'll need to make sure we get the correct
Participant for upcoming Zoom and Discord features.

This changes the behavior and introduces a new:
`find_participants_by_contact_id(...)` method that can be used if you
want to get the list of Participants for a Contact. We'll use that
in the RegisterUserAccount service which is the only place that calls
find_participant() with a nil program_id.

* Refactor RegisterUserAccount to use new find_participants_by_contact_id() method

This keeps the old behavior by just arbitrarily grabbing the first one returned.
I queued up a BUG here to fix the registration so it behaves in a more well-defined
manner for folks with multiple Participants: https://app.asana.com/0/1201131148207877/1201246645700107

* Hardcode a test program in the Discord signup-flow.

It's under active development and we still haven't decided where the
page will live so we don't know how we'll determine the Program that the
sign-up is for at the moment. Task to fix:
https://app.asana.com/0/1201131148207877/1201246645700111

* For Discord demo don't hardcode Program ID.

It won't work b/n dev and prod. Instead use the buggy approach
and just grab the first Enrolled Participant for now until we fix.